+++ Application/Application/DomainModel.h
@@ -11,7 +11,7 @@ public:
 	void AddShapeLogic(std::shared_ptr<ShapeLogic> & shapeLogic);
 	void RemoveShapeLogic(int pos);
 	Canvas& GetCanvas();
-	std::shared_ptr<ShapeLogic>& GetShape(int pos);
+	std::shared_ptr<ShapeLogic>& GetShape(size_t pos);
 	std::vector<std::shared_ptr<ShapeLogic>>& GetShapes();
 private:
 	std::vector<std::shared_ptr<ShapeLogic>> m_shapesLogic;
